requirements.txt
urllib3
requests
logging2
pyyaml
validators